requirements.txt
# Property Listing AI System - Iteration 1

# LangChain & LangGraph
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20
langchain-tavily>=0.0.1

# Environment variables
python-dotenv>=1.0.0

# UI
gradio>=4.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

